+++ src/main/java/com/mraof/minestuck/data/recipe/MinestuckCombinationsProvider.java
@@ -545,9 +545,9 @@ public class MinestuckCombinationsProvider extends RecipeProvider
 		CombinationRecipeBuilder.of(MSItems.SCARLET_ZILLYHOO.get()).input(MSItems.ZILLYHOO_HAMMER.get()).and().input(MSItems.FROG.get()).build(consumer);
 		CombinationRecipeBuilder.of(MSItems.MWRTHWL.get()).input(MSItems.REGI_HAMMER.get()).or().input(MSItems.CUEBALL.get()).build(consumer);
 		
-		CombinationRecipeBuilder.of(MSItems.STAFF_HANDLE.get()).input(MSItems.MINI_WIZARD_STATUE.get()).and().input(MSItems.CANE.get()).build(consumer);
-		CombinationRecipeBuilder.of(MSItems.WATER_STAFF.get()).input(MSItems.STAFF_HANDLE.get()).and().input(Items.HEART_OF_THE_SEA).build(consumer);
-		CombinationRecipeBuilder.of(MSItems.FIRE_STAFF.get()).input(MSItems.STAFF_HANDLE.get()).and().input(Items.MAGMA_CREAM).build(consumer);
+		CombinationRecipeBuilder.of(MSItems.WIZARD_STAFF.get()).input(MSItems.MINI_WIZARD_STATUE.get()).and().input(MSItems.CANE.get()).build(consumer);
+		CombinationRecipeBuilder.of(MSItems.WATER_STAFF.get()).input(MSItems.WIZARD_STAFF.get()).and().input(Items.HEART_OF_THE_SEA).build(consumer);
+		CombinationRecipeBuilder.of(MSItems.FIRE_STAFF.get()).input(MSItems.WIZARD_STAFF.get()).and().input(Items.MAGMA_CREAM).build(consumer);
 		CombinationRecipeBuilder.of(MSItems.PRIME_STAFF.get()).input(MSItems.BLACK_KINGS_SCEPTER.get()).and().input(MSItems.SBAHJ_POSTER.get()).build(consumer);
 		
 		CombinationRecipeBuilder.of(MSItems.VAUDEVILLE_HOOK.get()).namedInput(MSItems.CANE.get()).and().input(Items.FISHING_ROD).build(consumer);