+++ CHANGELOG.md
@@ -16,9 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Expanded number of Minecraft/Forge item/block tags that Minestuck items/blocks are in
 - Anthvil block
 - Power hub block
-- Fire and Water Staffs
-- Staff Handle
-- Rubik's Mace Weapon
+- Fire, Water, and Wizard Staffs
+- Rubik's Mace
 
 ### Changed